+++ Entities/Results/Parking/ParkingInfo.cs
@@ -9,7 +9,7 @@ namespace IPSCM.Entities.Results.Parking
     public class ParkingInfo
     {
         public UInt32 RecordId { get; set; }
-        public UInt64? Money { get; set; }
+        public Decimal? Money { get; set; }
         public String PhoneNumber { get; set; }
         public UInt32? UserId { get; set; }
     }
