+++ Protocol/EventArgs/LeavingEventArgs.cs
@@ -10,7 +10,7 @@ namespace IPSCM.Protocol.EventArgs
     public class LeavingEventArgs : HttpDataEventArgs
     {
         public LeavingEventArgs(HttpListenerRequest request, HttpListenerResponse response, String plateNumber,
-            DateTime outTime, Byte[] outImg, UInt32 copeMoney, UInt32 actualMoney, UInt64 ticketId)
+            DateTime outTime, Byte[] outImg, Decimal copeMoney, Decimal actualMoney, UInt32 ticketId)
             : base(request, response)
         {
             this.PlateNumber = plateNumber;
@@ -24,8 +24,8 @@ namespace IPSCM.Protocol.EventArgs
         public String PlateNumber { get; private set; }
         public DateTime OutTime { get; private set; }
         public Byte[] OutImg { get; private set; }
-        public UInt32 copeMoney { get; private set; }
-        public UInt32 actualMoney { get; private set; }
-        public UInt64 TicketId { get; private set; }
+        public Decimal copeMoney { get; private set; }
+        public Decimal actualMoney { get; private set; }
+        public UInt32 TicketId { get; private set; }
     }
 }
